requirements.txt
fastapi
uvicorn
joblib
numpy
Pillow
opencv-python-headless
scikit-image
scikit-learn==1.4.2
python-multipart
pydantic
ultralytics